(LedgerStore) Move metric-related functions to blockstore_metric.rs
